Fix scheduled events failing with unsupported event error (#5)
Previously the versioning action defaulted to a local alpha version for any unknown event types. Now that event types are explicitly checked an error is thrown if an unknown type is encountered.
Handling for the 'schedule' event was missing, which caused failures for providers with cron jobs.
This change adds the necessary handling for 'schedule' events and constructs a local development version for those like before.