Check the PR head ref, not base ref
The base of a PR is where it's merging into (e.g. "main"), but we want to know the name of the branch that is *being* merged. This is the 'head'.

We can't use the `context.ref` because this is the merge target e.g. `refs/pull/524/merge`

See example from debug logs:
[`head.ref`](https://github.com/pulumi/pulumi-oci/actions/runs/9719486666/job/26934553104#step:3:286): `upgrade-terraform-provider-oci-to-v6.0.0-major`
[`base.ref`](https://github.com/pulumi/pulumi-oci/actions/runs/9719486666/job/26934553104#step:3:133): `main`
[`ref`](https://github.com/pulumi/pulumi-oci/actions/runs/9719486666/job/26934553104#step:3:607): `refs/pull/524/merge`

Made branch functions a little more robust against missing values.